Fix emitted characteristics JSON to be compliant
Currently the characteristics JSON does not fully comply with the HomeKit
protocol when a characteristic returns an error, causing iOS to error
out when communicating with the bridge.

A characteristic may return an error via ValueRequestFunc, but calling
`GET /accessories` must still provide a valid "value" in the JSON.

Additionally, when a 207 Multi-Status response is returned because of
this, characteristics that are readable must contain a "value" _and_ an
explicit "status: 0", whereas characteristics with errors must contain a
non-zero "status" and must not contain a "value".

Fixes #30.