Implement WriteResponse characteristics
Some control points require that the response be returned after a write,
as opposed to write-then-read. If the SetValueRequestFunc returns a
value and has the WriteResponse permission, then the response is relayed
back to the client on return. Status code must be explicitly returned in
this case, via a MultiStatus return.